Move synchronizer from iguazio package to client package
