+++ pkg/platform/abstract/project/external/leader/client/synchronizer.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package iguazio
+package client
 
 import (
 	"context"
